[GTK][WPE] Make sure we just use the triggerMemoryPressureEvent with information from the UIProcess for the MemoryPressureHandler https://bugs.webkit.org/show_bug.cgi?id=304428
Reviewed by NOBODY (OOPS!).

We are using the m_measurementTimer and the triggerMemoryPressureEvent signaled by
MemoryPressureMonitor at the same time, this is not correct because both have the same
goal and the one in the timer uses a suboptimal implementation to check memory usage
that causes stops in the main loop of the WebProcess. We want to first use just the one
in the MemoryPressureMonitor and later verify how to unify and make it more similar to
other platforms.

* Source/WTF/wtf/MemoryPressureHandler.cpp:
(WTF::MemoryPressureHandler::setShouldUsePeriodicMemoryMonitor):